fixup! build: add Bazel build system
