+++ jest.config.js
@@ -3,5 +3,5 @@ module.exports = {
   "moduleFileExtensions": [
     "js"
   ],
-  "snapshotResolver": "<rootDir>/snapshotResolver.js",
+  "snapshotResolver": "./snapshotResolver.js",
 };
