+++ packages/BUILD.bazel
@@ -2,5 +2,4 @@ package(default_visibility = ["//visibility:public"])
 
 exports_files([
     "tsconfig-build.json",
-    "README.md",
 ])
