--- packages/README.md
@@ -1,6 +1,0 @@
-AWS CDK8s
-=======
-
-The sources for this package are in the main [AWS CDK8S](https://github.com/awslabs/cdk8s) repo. Please file issues and pull requests against that repo.
-
-License: MIT
