--- packages/cdk8s/jest.config.js
@@ -1,7 +1,0 @@
-module.exports = {
-  "collectCoverage": true,
-  "moduleFileExtensions": [
-    "js"
-  ],
-  "snapshotResolver": "<rootDir>/snapshotResolver.js",
-};
