+++ tools/BUILD.bazel
@@ -14,6 +14,5 @@ nodejs_binary(
 )
 
 exports_files([
-    "jest.config.js",
     "package-json/package.json",
 ])
